+++ tests/test_stl.cpp
@@ -93,16 +93,16 @@ NB_MODULE(test_stl_ext, m) {
     nb::class_<Movable>(m, "Movable")
         .def(nb::init<>())
         .def(nb::init<int>())
-        .def_readwrite("value", &Movable::value);
+        .def_rw("value", &Movable::value);
 
     nb::class_<Copyable>(m, "Copyable")
         .def(nb::init<>())
         .def(nb::init<int>())
-        .def_readwrite("value", &Copyable::value);
+        .def_rw("value", &Copyable::value);
 
     nb::class_<StructWithReadonlyMap>(m, "StructWithReadonlyMap")
         .def(nb::init<>())
-        .def_readonly("map", &StructWithReadonlyMap::map);
+        .def_ro("map", &StructWithReadonlyMap::map);
 
     // ----- test01-test12 ------
 
@@ -236,8 +236,8 @@ NB_MODULE(test_stl_ext, m) {
 
     nb::class_<FuncWrapper>(m, "FuncWrapper", nb::type_slots(slots))
         .def(nb::init<>())
-        .def_readwrite("f", &FuncWrapper::f)
-        .def_readonly_static("alive", &FuncWrapper::alive);
+        .def_rw("f", &FuncWrapper::f)
+        .def_ro_static("alive", &FuncWrapper::alive);
 
     // ----- test35 ------
     m.def("identity_string", [](std::string& x) { return x; });
@@ -405,5 +405,5 @@ NB_MODULE(test_stl_ext, m) {
 
     nb::class_<ClassWithMovableField>(m, "ClassWithMovableField")
         .def(nb::init<>())
-        .def_readwrite("movable", &ClassWithMovableField::movable);
+        .def_rw("movable", &ClassWithMovableField::movable);
 }